Store alpha as float8 to make it consistent
We read and validate the alpha values as float8/double, but the structs
used just float4 to store the value. That may result in unexpected
issues due to rounding errors etc. So store it as float8, to keep it
consistent.

This also affects the the regression test output, because the float8
value may be a bit different, which affects the binning (although only
very slightly).