Remove the bucket index calculation
The ddsketch paper assumes the values are positive integers, which
means the log(value) returns a non-negative value. The code however did
log(1.0 + value) to handle all values >= 0.0, but this difference from
the paper was arbitrary and might have undesirable consequences. This
switches to the formula from the paper, with additional restrictions on
the input values to reject values < 1.0.

This also affects the binning, so the expected regression test output
needs to be adopted too.